apps/ml-agent-fastapi/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.0.0
pandas>=2.0.0
prophet>=1.1.5
scikit-learn>=1.4.0
langchain>=0.2.0
langchain-mongodb>=0.1.0
langchain-openai>=0.1.0
pymongo>=4.0.0
python-dotenv>=1.0.0
opentelemetry-api>=1.23.0
opentelemetry-sdk>=1.23.0
opentelemetry-exporter-otlp>=1.23.0
opentelemetry-instrumentation-fastapi>=0.44b0
temporalio>=1.5.0
azure-identity>=1.15.0
azure-keyvault-secrets>=4.8.0
